Изменил структуру таблиц заказа и прихода, изменил поведения в контроллерах и моделях

+++ modules/api/controllers/OrderController.php
@@ -11,6 +11,7 @@ namespace app\modules\api\controllers;
 use app\modules\trade\models\Order;
 use app\modules\trade\models\OrderItem;
 use Yii;
+use yii\db\Exception;
 use yii\rest\ActiveController;
 
 class OrderController extends ActiveController
@@ -23,11 +24,11 @@ class OrderController extends ActiveController
 
         $document = Order::findOne($requestParams['Order']['id']);
         $documentTable = json_decode($requestParams['documentTable']);
-        $documentTableInitial = OrderItem::find()->where(['order_id' => $requestParams['Order']['id']])->with('good')->all();
+        $documentTableInitial = OrderItem::find()->where(['order_id' => $requestParams['Order']['id']])->all();
 
         foreach ($documentTableInitial as $index => $item) {
             foreach ($documentTable as $value) {
-                if ($item->good->article === $value->article) {
+                if ($item['good_id'] == $value->good_id) {
                     unset($documentTableInitial[$index]);
                 }
             }
@@ -35,6 +36,12 @@ class OrderController extends ActiveController
 
         $transaction = Order::getDb()->beginTransaction();
         try {
+
+            if (count($documentTable) === 0) {
+                throw new Exception('ошибка: ', ['error' =>
+                                                     ['Ошибка: В документе должна быть хотя бы одна строка']
+                ]);
+            }
 
             if (!isset($document)) {
                 $document = new Order();
@@ -44,31 +51,43 @@ class OrderController extends ActiveController
             $document->date = $requestParams['Order']['date'];
             $document->status = $requestParams['Order']['status'];
             $document->payment = $requestParams['Order']['payment'];
-            $document->save();
+            if ($document->save()) {
+                foreach ($documentTable as $item) {
+                    $tableItem = OrderItem::find()->where(['order_id' => $document->id, 'good_id' => $item->good_id])->one();
+                    if (!isset($tableItem)) {
+                        $tableItem = new OrderItem();
+                        $tableItem->order_id = $document->id;
+                        $tableItem->good_id = $item->good_id;
+                    }
 
-            foreach ($documentTable as $item) {
-                $tableItem = OrderItem::findOne($item->id);
-                if (!isset($tableItem)) {
-                    $tableItem = new OrderItem();
-                    $tableItem->order_id = $document->id;
-                    $tableItem->good_id = $item->good_id;
+                    $tableItem->qty = $item->qty;
+                    $tableItem->price = $item->price;
+                    $tableItem->save();
                 }
 
-                $tableItem->qty = $item->qty;
-                $tableItem->price = $item->price;
-                $tableItem->save();
+                foreach ($documentTableInitial as $item) {
+                    $item->delete();
+                }
+
+            } else {
+                throw new Exception('ошибка', $document->errors);
             }
 
-            foreach ($documentTableInitial as $item) {
-                $item->delete();
-            }
 
             $transaction->commit();
-            return true;
+            if (!$document->hasErrors() && strpos(Yii::$app->request->referrer, 'create')) {
+                return ['status' => true, 'id' => $document->id];
+            } else {
+                return ['status' => true];
+            }
+
+        } catch (\Exception $e) {
+            $transaction->rollBack();
+            return $e;
         } catch (\Throwable $e) {
             $transaction->rollBack();
             throw $e;
-            return false;
+            return ['status' => false];
         }
     }
 }
